backend/requirements.txt
# --- Core Framework ---
fastapi==0.110.0
uvicorn[standard]==0.29.0

# --- Environment & Config ---
python-dotenv==1.0.1

# --- Azure SDKs ---
azure-identity==1.16.0
azure-core==1.30.1
azure-search-documents==11.4.0
openai==1.14.3

# --- LLM / Tokenization ---
tiktoken==0.6.0

# --- HTTP / Utils ---
requests==2.31.0

# --- Optional (for future observability) ---
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
